pyproject.toml
[project]
name = "463-sql-llm-udemy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.3",
    "langchain-core>=1.2.6",
    "langchain-experimental>=0.4.1",
    "langchain-openai>=1.1.7",
    "openai>=2.14.0",
    "pandas>=2.3.3",
    "pyodbc>=5.3.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "streamlit>=1.52.2",
    "tabulate>=0.9.0",
]
